requirements.txt
python-telegram-bot==24.2
